KAFKA-989 Port EEP specific changes to Kafka 3.6.1, Part 1: kafka-clients & kafka-eventstreams (2)
1. Added mapr-eco-config.sh file where added mapr-eco-tools and slf4j-reload4j jars to "$(mapr clientclasspath --kafka)"
2. Made testSrcJar tasks depend on processMessages tasks to avoid gradle build error
3. For eventstreams test tasks, added maprdb dependency and java version 11 so they can compile
4. Duplicated KafkaConsumer.kafkaConsumerMetrics initialization in maybeMaprInitialize to avoid NPEs
5. Throw from Marlin when calling consumer.position(partition) for unassigned partition (because KafkaConsumer doesn't know what partitions were assigned by subscribe())
6. Until KafkaConsumer#currentLag() API is not implemented, it will return OptionalLong.empty()
7. MaprKafkaUtils: removed dead code, tuned isMapr and maybeWrapDefaultStream methods, added maybeTrimTopic and listAllTopics
8. Topic.validate now allows '/' and ':'
9. Until Admin#deleteRecords() is not implemented, it will return empty result
10. Fixed Duration->long convertion in MarlinListener#poll(Duration)
11. auto.commit.interval.ms is int, not long

+++ eventstreams/src/main/java/com/mapr/kafka/eventstreams/impl/listener/MarlinListenerImpl.java
@@ -200,7 +200,7 @@ public class MarlinListenerImpl extends MarlinJniListener {
                               this.rpcTimeoutMs,
                               hardMount,
                               config.getBoolean(mConfDef.getAutoCommitEnabled()),
-                              config.getLong(mConfDef.getAutoCommitInterval()),
+                              config.getInt(mConfDef.getAutoCommitInterval()),
                               null /* autoCommitCb, applicable for V10 */,
                               config.getLong(mConfDef.getMetadataMaxAge()),
                               config.getInt(mConfDef.getFetchMsgMaxBytesPerPartition()),
